Removed double filt.filt df name and made formating changes.
There remain issues in here regarding the use of the applyRunsFilter. The issue (on github) of that function returning a data frame needs to be resolved before dealing with the issues here (which are redundancies in the code, since the structure of df.alltags.filt is already a data frame).